Long int change

+++ main.h
@@ -40,6 +40,6 @@ int restriction_percentage(const char *str);
 int (*match_case(const char *))(va_list);
 int convert_base(int base, long int number, int band);
 /*Task to print number*/
-int print_number(int n);
+int print_number(long int n);
 
 #endif
